requirements.txt
pdfminer.six==20240706
openai>=1.52.1

